Move cache code to an extension
